Complete method and tests for admin discounted revenue

+++ db/migrate/10.rb
@@ -1,5 +1,5 @@
 class ChangeDataTypeForPercentage < ActiveRecord::Migration[5.2]
   def change
-    change_column(:discounts, :percentage, :integer)
+    change_column(:discounts, :percentage, :decimal)
   end
 end
